pyproject.toml
[project]
name = "life-planner-lite"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "matplotlib>=3.10.8",
    "plotly>=6.6.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]
